+++ packages/perseus/src/prompt-types.ts
@@ -6,7 +6,7 @@ import type {ExpressionPromptJSON} from "./widget-ai-utils/expression/expression
 import type {GradedGroupPromptJSON} from "./widget-ai-utils/graded-group/graded-group-ai-utils";
 import type {GradedGroupSetPromptJSON} from "./widget-ai-utils/graded-group-set/graded-group-set-ai-utils";
 import type {GrapherPromptJSON} from "./widget-ai-utils/grapher/grapher-ai-utils";
-import type {GroupPromptJSON} from "./widget-ai-utils/group/prompt-utils";
+import type {GroupPromptJSON} from "./widget-ai-utils/group/group-ai-utils";
 import type {ImagePromptJSON} from "./widget-ai-utils/image/prompt-utils";
 import type {InputNumberPromptJSON} from "./widget-ai-utils/input-number/prompt-utils";
 import type {LabelImagePromptJSON} from "./widget-ai-utils/label-image/prompt-utils";
