+++ packages/perseus/src/widget-ai-utils/group/group-ai-utils.ts
