+++ packages/perseus/src/widgets/group/group.tsx
@@ -6,7 +6,7 @@ import {PerseusI18nContext} from "../../components/i18n-context";
 import * as Changeable from "../../mixins/changeable";
 import {ApiOptions} from "../../perseus-api";
 import Renderer from "../../renderer";
-import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/group/prompt-utils";
+import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/group/group-ai-utils";
 
 import type {PerseusGroupWidgetOptions} from "../../perseus-types";
 import type {
@@ -18,7 +18,7 @@ import type {
     WidgetProps,
 } from "../../types";
 import type {PerseusGroupRubric, UserInputArray} from "../../validation.types";
-import type {GroupPromptJSON} from "../../widget-ai-utils/group/prompt-utils";
+import type {GroupPromptJSON} from "../../widget-ai-utils/group/group-ai-utils";
 
 type RenderProps = PerseusGroupWidgetOptions; // exports has no 'transform'
 type Props = WidgetProps<RenderProps, PerseusGroupRubric>;
